+++ packages/amm/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sushiswap/amm
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [[`c6d6cc5db`](https://github.com/sushiswap/sushiswap/commit/c6d6cc5db4cc614f3931ee3a325547967c86c51a)]:
+  - @sushiswap/chain@0.1.2
+  - @sushiswap/trident-core@1.0.5
+  - @sushiswap/currency@0.0.6
+  - @sushiswap/v2-sdk@0.0.5
+  - @sushiswap/tines@1.0.6
+
 ## 0.1.8
 
 ### Patch Changes